Moved Promiseinfo to correct folder
