+++ packages/browser-wallet-api/src/promiseInfo.ts
